requirements.txt
Flask==2.2.3
google-auth==2.19.0
google-auth-oauthlib==0.7.2
google-auth-httplib2==0.1.0
google-api-python-client==2.80.0
